check if reads were successful
Reviewed-on: https://gerrit.libreoffice.org/14631
Tested-by: Caolán McNamara <caolanm@redhat.com>
(cherry picked from commit f974db5d89eacf0c23e303c22c62972014e9db16)

Conflicts:
	hwpfilter/source/hiodev.cxx
	hwpfilter/source/hiodev.h
	hwpfilter/source/hwpfile.cxx

Reviewed-on: https://gerrit.libreoffice.org/14654
Tested-by: Michael Stahl <mstahl@redhat.com>
(cherry picked from commit f2d49715c176c80c4b0fa3a7799d610eb5afec88)
(cherry picked from commit 49c4b067f5c209b40d06804c2399fb1706b92282)

Conflicts:
	hwpfilter/source/drawdef.h
	hwpfilter/source/hiodev.h

+++ hwpfilter/source/hbox.h
@@ -71,7 +71,7 @@ struct HBox
  */
 struct SkipData: public HBox
 {
-    ulong data_block_len;
+    uint data_block_len;
     hchar dummy;
     char  *data_block;
 
@@ -631,7 +631,7 @@ struct Picture: public FBox
  * follow_block_size is the size information of the Drawing object of hwp.
  * It's value is greater than 0 if the pictype is PICTYPE_DRAW.
  */
-    ulong     follow_block_size;                  /* 추가정보 길이. */
+    uint      follow_block_size;                  /* 추가정보 길이. */
     short     dummy1;                             // to not change structure size */
     short     dummy2;                             // to not change structure size */
     uchar     reserved1;
